+++ sui_core/src/lib.rs
@@ -9,4 +9,5 @@ pub mod authority_client;
 pub mod authority_server;
 pub mod client;
 pub mod consensus_handler;
+pub mod execution_engine;
 pub mod safe_client;
